+++ packages/terra-dynamic-grid/CHANGELOG.md
@@ -3,12 +3,12 @@ ChangeLog
 
 Unreleased
 ----------
-### Changed
-* Import for package version from package.json added to doc page
 
-Unreleased
+3.15.0 - (August 21, 2019)
 ------------------
 ### Changed
+* Import for package version from package.json added to doc page
+* Added comments for ESlint v6
 * Cleaned up imports in examples and test files
 
 3.14.0 - (August 14, 2019)