+++ packages/terra-progress-bar/CHANGELOG.md
@@ -3,12 +3,11 @@ Changelog
 
 Unreleased
 ----------
-### Changed
-* Import for package version from package.json added to doc page
 
-Unreleased
+4.12.0 - (August 21, 2019)
 ------------------
 ### Changed
+* Import for package version from package.json added to doc page
 * Cleaned up imports in examples and test files
 
 4.11.0 - (August 14, 2019)